[BC] Fix phpcs SlevomatCodingStandard.Classes
