+++ tests/GeneratedHydratorTestAsset/Base.php
